requirements.txt
numpy
pandas
scikit-learn
scipy
matplotlib
seaborn
nltk
spacy
imbalanced-learn
torch
transformers
accelerate
datasets
mlflow
hydra-core
PyYAML
omegaconf
pydantic
fastapi
uvicorn
jupyter
notebook
tqdm
rich
pytest
black
ruff
mypy
pre-commit
statsmodels
dvc
umap-learn
wordcloud
jinja2
kaggle
requests
gitpython
transformers-interpret
captum
pip-tools
